Nova funcionalidade: Cadastro de dispostivo
Agora é possivel cadastrar um dispositivo em uma residencia

Foi um script em java script, alterada a base de dados, controllers e criadas novas views. Domains e Services também foram alterados e implementados

+++ lumiere-mvc/src/main/java/com/lumiere/boot/dao/TipoDispositivoDao.java
@@ -7,13 +7,13 @@ import com.lumiere.boot.domain.TipoDispositivo;
 
 public interface TipoDispositivoDao {
 
-void save(TipoDispositivo tipodispositivo);
+	void save(TipoDispositivo tipodispositivo);
 	
 	void update(TipoDispositivo tipodispositivo);
 	
-	void delete(Long id);
+	void delete(Integer id);
 	
-	TipoDispositivo findById(Long id);
+	TipoDispositivo findById(Integer id);
 	
 	List<TipoDispositivo> findAll();
 }
